requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
flask
llama-cpp-python
pypdf